requirements.txt
asgiref
Django
sqlparse
tzdata
pillow
django-taggit